# opt: reduce index partition locality allocations