--- a/pkg/sql/BUILD.bazel
+++ b/pkg/sql/BUILD.bazel
@@ -378,6 +378,7 @@ go_library(
         "//pkg/sql/opt/memo",
         "//pkg/sql/opt/norm",
         "//pkg/sql/opt/optbuilder",
+        "//pkg/sql/opt/partition",
         "//pkg/sql/opt/xform",
         "//pkg/sql/optionalnodeliveness",
         "//pkg/sql/paramparse",