#### opt: return non-pointer from HasMixOfLocalAndRemotePartitions

Release note: None

#### opt: move TableMeta.AddPartialIndexPredicate next to TableMeta.PartialIndexPredicate

Release note: None

#### opt: reduce index partition locality allocations

Prior to this commit, a map was allocated in the table metadata to store
`PrefixSorter`s for the indexes of the table. This map was allocated
even if no indexes were partitioned. This commit introduces several
changes that reduce the number of allocations:

  1. The map has been replaced with a slice, which will require only one
     allocation because it will never grow.
  2. The slice is only allocated if there is at least one index with a
     mix of local and remote partitions.
  3. The slice is eagerly added to the table metadata instead of lazily.
     The prefix sorters are accessed in very common rules like
     GenerateConstrainedScans, so there was no benefit in lazily
     creating them.

Release justification: Minor change that fixes micro-benchmark
regressions.

Release note: None