@mgartner 2022-08-19
Benchmark results show up to a 6% reduction in planning latency and up to a 67% reduction in allocated memory:

`BenchmarkPhases` (with parsing and optbuild phases omitted - they showed no change):

```
name                                                                     old time/op    new time/op    delta
Phases/kv-read/Simple/Explore-24                                           45.5µs ± 1%    45.5µs ± 1%     ~     (p=0.548 n=5+5)
Phases/kv-read/Simple/ExecBuild-24                                         48.1µs ± 1%    47.8µs ± 1%     ~     (p=0.151 n=5+5)
Phases/kv-read/Prepared/ExecBuild-24                                       1.29µs ± 0%    1.35µs ± 0%   +4.50%  (p=0.008 n=5+5)
Phases/kv-read-const/Simple/Explore-24                                     45.9µs ± 0%    45.9µs ± 0%     ~     (p=0.460 n=5+5)
Phases/kv-read-const/Simple/ExecBuild-24                                   48.3µs ± 0%    48.3µs ± 0%     ~     (p=0.151 n=5+5)
Phases/kv-read-const/Prepared/ExecBuild-24                                 1.09µs ± 0%    1.09µs ± 0%     ~     (p=0.397 n=5+5)
Phases/tpcc-new-order/Simple/Explore-24                                     133µs ± 0%     133µs ± 1%     ~     (p=0.841 n=5+5)
Phases/tpcc-new-order/Simple/ExecBuild-24                                   137µs ± 1%     137µs ± 0%     ~     (p=0.548 n=5+5)
Phases/tpcc-new-order/Prepared/ExecBuild-24                                1.66µs ± 0%    1.72µs ± 0%   +3.29%  (p=0.008 n=5+5)
Phases/tpcc-delivery/Simple/Explore-24                                      104µs ± 0%     104µs ± 1%     ~     (p=0.841 n=5+5)
Phases/tpcc-delivery/Simple/ExecBuild-24                                    108µs ± 0%     108µs ± 0%   +0.39%  (p=0.032 n=5+5)
Phases/tpcc-delivery/Prepared/AssignPlaceholdersNoNorm-24                  13.5µs ± 0%    13.7µs ± 1%   +1.14%  (p=0.008 n=5+5)
Phases/tpcc-delivery/Prepared/AssignPlaceholdersNorm-24                    13.5µs ± 0%    13.7µs ± 1%   +1.54%  (p=0.008 n=5+5)
Phases/tpcc-delivery/Prepared/Explore-24                                   41.9µs ± 0%    41.3µs ± 0%   -1.28%  (p=0.008 n=5+5)
Phases/tpcc-delivery/Prepared/ExecBuild-24                                 44.4µs ± 1%    43.7µs ± 1%   -1.63%  (p=0.008 n=5+5)
Phases/tpcc-stock-level/Simple/Explore-24                                   507µs ± 1%     509µs ± 0%     ~     (p=0.151 n=5+5)
Phases/tpcc-stock-level/Simple/ExecBuild-24                                 522µs ± 0%     526µs ± 1%   +0.70%  (p=0.016 n=5+5)
Phases/tpcc-stock-level/Prepared/AssignPlaceholdersNoNorm-24               58.5µs ± 1%    58.2µs ± 0%     ~     (p=0.151 n=5+5)
Phases/tpcc-stock-level/Prepared/AssignPlaceholdersNorm-24                 60.0µs ± 0%    59.7µs ± 0%   -0.61%  (p=0.016 n=5+5)
Phases/tpcc-stock-level/Prepared/Explore-24                                 199µs ± 0%     198µs ± 0%   -0.56%  (p=0.008 n=5+5)
Phases/tpcc-stock-level/Prepared/ExecBuild-24                               215µs ± 0%     214µs ± 0%   -0.54%  (p=0.016 n=5+5)
Phases/many-columns-and-indexes-a/Simple/Explore-24                         505µs ± 0%     476µs ± 0%   -5.79%  (p=0.008 n=5+5)
Phases/many-columns-and-indexes-a/Simple/ExecBuild-24                       511µs ± 0%     483µs ± 1%   -5.45%  (p=0.008 n=5+5)
Phases/many-columns-and-indexes-a/Prepared/AssignPlaceholdersNoNorm-24      273µs ± 0%     274µs ± 0%   +0.33%  (p=0.008 n=5+5)
Phases/many-columns-and-indexes-a/Prepared/AssignPlaceholdersNorm-24        274µs ± 1%     274µs ± 0%     ~     (p=0.310 n=5+5)
Phases/many-columns-and-indexes-a/Prepared/Explore-24                       460µs ± 0%     431µs ± 0%   -6.27%  (p=0.008 n=5+5)
Phases/many-columns-and-indexes-a/Prepared/ExecBuild-24                     464µs ± 0%     436µs ± 0%   -6.19%  (p=0.008 n=5+5)
Phases/many-columns-and-indexes-b/Simple/Explore-24                         543µs ± 0%     514µs ± 0%   -5.43%  (p=0.008 n=5+5)
Phases/many-columns-and-indexes-b/Simple/ExecBuild-24                       552µs ± 0%     523µs ± 0%   -5.22%  (p=0.008 n=5+5)
Phases/many-columns-and-indexes-b/Prepared/AssignPlaceholdersNoNorm-24      280µs ± 0%     281µs ± 0%   +0.38%  (p=0.008 n=5+5)
Phases/many-columns-and-indexes-b/Prepared/AssignPlaceholdersNorm-24        280µs ± 0%     281µs ± 0%   +0.51%  (p=0.008 n=5+5)
Phases/many-columns-and-indexes-b/Prepared/Explore-24                       472µs ± 0%     442µs ± 0%   -6.29%  (p=0.008 n=5+5)
Phases/many-columns-and-indexes-b/Prepared/ExecBuild-24                     478µs ± 0%     448µs ± 0%   -6.28%  (p=0.008 n=5+5)
Phases/many-columns-and-indexes-c/Simple/Explore-24                        48.8ms ± 0%    48.7ms ± 0%     ~     (p=0.095 n=5+5)
Phases/many-columns-and-indexes-c/Simple/ExecBuild-24                      49.1ms ± 0%    48.9ms ± 0%   -0.42%  (p=0.016 n=5+5)
Phases/many-columns-and-indexes-c/Prepared/AssignPlaceholdersNoNorm-24     10.6ms ± 0%    10.6ms ± 1%   +0.79%  (p=0.008 n=5+5)
Phases/many-columns-and-indexes-c/Prepared/AssignPlaceholdersNorm-24       10.6ms ± 0%    10.6ms ± 0%   +0.53%  (p=0.008 n=5+5)
Phases/many-columns-and-indexes-c/Prepared/Explore-24                      46.5ms ± 0%    46.5ms ± 0%     ~     (p=0.222 n=5+5)
Phases/many-columns-and-indexes-c/Prepared/ExecBuild-24                    46.8ms ± 0%    46.8ms ± 0%     ~     (p=0.548 n=5+5)
Phases/batch-insert-one/Simple/Explore-24                                   110µs ± 1%     110µs ± 0%     ~     (p=0.690 n=5+5)
Phases/batch-insert-one/Simple/ExecBuild-24                                 120µs ± 0%     121µs ± 1%     ~     (p=0.056 n=5+5)
Phases/batch-insert-one/Prepared/ExecBuild-24                              5.17µs ± 0%    5.17µs ± 1%     ~     (p=0.452 n=5+5)
Phases/batch-insert-many/Simple/Explore-24                                 7.78ms ± 1%    7.82ms ± 1%     ~     (p=0.548 n=5+5)
Phases/batch-insert-many/Simple/ExecBuild-24                               7.84ms ± 1%    7.86ms ± 1%     ~     (p=0.690 n=5+5)
Phases/batch-insert-many/Prepared/ExecBuild-24                             78.3µs ± 0%    75.1µs ± 1%   -4.00%  (p=0.008 n=5+5)
Phases/ored-preds-100/Simple/Explore-24                                    3.32ms ± 0%    3.32ms ± 0%     ~     (p=0.548 n=5+5)
Phases/ored-preds-100/Simple/ExecBuild-24                                  3.43ms ± 0%    3.43ms ± 0%     ~     (p=0.548 n=5+5)
Phases/ored-preds-100/Prepared/ExecBuild-24                                 104µs ± 0%     103µs ± 0%   -0.50%  (p=0.008 n=5+5)
Phases/ored-preds-using-params-100/Simple/Explore-24                       3.14ms ± 0%    3.13ms ± 0%   -0.41%  (p=0.032 n=5+5)
Phases/ored-preds-using-params-100/Simple/ExecBuild-24                     3.24ms ± 0%    3.26ms ± 1%     ~     (p=0.151 n=5+5)
Phases/ored-preds-using-params-100/Prepared/AssignPlaceholdersNoNorm-24     568µs ± 0%     568µs ± 1%     ~     (p=1.000 n=5+5)
Phases/ored-preds-using-params-100/Prepared/AssignPlaceholdersNorm-24       567µs ± 0%     565µs ± 0%     ~     (p=0.310 n=5+5)
Phases/ored-preds-using-params-100/Prepared/Explore-24                     1.43ms ± 0%    1.43ms ± 1%     ~     (p=0.151 n=5+5)
Phases/ored-preds-using-params-100/Prepared/ExecBuild-24                   1.56ms ± 1%    1.54ms ± 0%   -0.76%  (p=0.008 n=5+5)

name                                                                     old alloc/op   new alloc/op   delta
Phases/kv-read/Simple/Explore-24                                           14.7kB ± 0%    14.5kB ± 0%   -1.31%  (p=0.008 n=5+5)
Phases/kv-read/Simple/ExecBuild-24                                         15.2kB ± 0%    15.0kB ± 0%   -1.27%  (p=0.000 n=4+5)
Phases/kv-read/Prepared/ExecBuild-24                                         704B ± 0%      704B ± 0%     ~     (all equal)
Phases/kv-read-const/Simple/Explore-24                                     14.8kB ± 0%    14.6kB ± 0%   -1.30%  (p=0.000 n=4+5)
Phases/kv-read-const/Simple/ExecBuild-24                                   15.3kB ± 0%    15.1kB ± 0%   -1.26%  (p=0.000 n=5+4)
Phases/kv-read-const/Prepared/ExecBuild-24                                   520B ± 0%      520B ± 0%     ~     (all equal)
Phases/tpcc-new-order/Simple/Explore-24                                    34.0kB ± 0%    33.8kB ± 0%   -0.56%  (p=0.008 n=5+5)
Phases/tpcc-new-order/Simple/ExecBuild-24                                  34.5kB ± 0%    34.3kB ± 0%   -0.56%  (p=0.008 n=5+5)
Phases/tpcc-new-order/Prepared/ExecBuild-24                                  768B ± 0%      768B ± 0%     ~     (all equal)
Phases/tpcc-delivery/Simple/Explore-24                                     26.3kB ± 0%    26.1kB ± 0%   -0.72%  (p=0.008 n=5+5)
Phases/tpcc-delivery/Simple/ExecBuild-24                                   26.7kB ± 0%    26.6kB ± 0%   -0.71%  (p=0.008 n=5+5)
Phases/tpcc-delivery/Prepared/AssignPlaceholdersNoNorm-24                  5.44kB ± 0%    5.44kB ± 0%     ~     (all equal)
Phases/tpcc-delivery/Prepared/AssignPlaceholdersNorm-24                    5.44kB ± 0%    5.44kB ± 0%     ~     (all equal)
Phases/tpcc-delivery/Prepared/Explore-24                                   13.1kB ± 0%    12.9kB ± 0%   -1.47%  (p=0.008 n=5+5)
Phases/tpcc-delivery/Prepared/ExecBuild-24                                 13.5kB ± 0%    13.3kB ± 0%   -1.42%  (p=0.008 n=5+5)
Phases/tpcc-stock-level/Simple/Explore-24                                   119kB ± 0%     118kB ± 0%   -0.31%  (p=0.008 n=5+5)
Phases/tpcc-stock-level/Simple/ExecBuild-24                                 123kB ± 0%     122kB ± 0%   -0.31%  (p=0.008 n=5+5)
Phases/tpcc-stock-level/Prepared/AssignPlaceholdersNoNorm-24               19.1kB ± 0%    19.1kB ± 0%     ~     (p=0.143 n=5+5)
Phases/tpcc-stock-level/Prepared/AssignPlaceholdersNorm-24                 18.8kB ± 0%    18.8kB ± 0%     ~     (p=1.000 n=5+4)
Phases/tpcc-stock-level/Prepared/Explore-24                                46.9kB ± 0%    46.5kB ± 0%   -0.83%  (p=0.008 n=5+5)
Phases/tpcc-stock-level/Prepared/ExecBuild-24                              50.7kB ± 0%    50.3kB ± 0%   -0.76%  (p=0.008 n=5+5)
Phases/many-columns-and-indexes-a/Simple/Explore-24                        44.2kB ± 0%    22.9kB ± 0%  -48.24%  (p=0.008 n=5+5)
Phases/many-columns-and-indexes-a/Simple/ExecBuild-24                      44.9kB ± 0%    23.6kB ± 0%  -47.51%  (p=0.008 n=5+5)
Phases/many-columns-and-indexes-a/Prepared/AssignPlaceholdersNoNorm-24     3.81kB ± 0%    3.81kB ± 0%     ~     (all equal)
Phases/many-columns-and-indexes-a/Prepared/AssignPlaceholdersNorm-24       3.81kB ± 0%    3.81kB ± 0%     ~     (all equal)
Phases/many-columns-and-indexes-a/Prepared/Explore-24                      31.7kB ± 0%    10.3kB ± 0%  -67.38%  (p=0.008 n=5+5)
Phases/many-columns-and-indexes-a/Prepared/ExecBuild-24                    32.3kB ± 0%    11.0kB ± 0%  -65.95%  (p=0.008 n=5+5)
Phases/many-columns-and-indexes-b/Simple/Explore-24                        51.1kB ± 0%    29.8kB ± 0%  -41.74%  (p=0.008 n=5+5)
Phases/many-columns-and-indexes-b/Simple/ExecBuild-24                      52.2kB ± 0%    30.8kB ± 0%  -40.92%  (p=0.008 n=5+5)
Phases/many-columns-and-indexes-b/Prepared/AssignPlaceholdersNoNorm-24     6.05kB ± 0%    6.05kB ± 0%     ~     (p=0.881 n=5+5)
Phases/many-columns-and-indexes-b/Prepared/AssignPlaceholdersNorm-24       6.05kB ± 0%    6.05kB ± 0%     ~     (p=1.000 n=5+5)
Phases/many-columns-and-indexes-b/Prepared/Explore-24                      33.5kB ± 0%    12.2kB ± 0%  -63.61%  (p=0.008 n=5+5)
Phases/many-columns-and-indexes-b/Prepared/ExecBuild-24                    34.6kB ± 0%    13.2kB ± 0%  -61.73%  (p=0.008 n=5+5)
Phases/many-columns-and-indexes-c/Simple/Explore-24                        26.3MB ± 0%    26.3MB ± 0%   -0.24%  (p=0.008 n=5+5)
Phases/many-columns-and-indexes-c/Simple/ExecBuild-24                      26.4MB ± 0%    26.4MB ± 0%   -0.24%  (p=0.008 n=5+5)
Phases/many-columns-and-indexes-c/Prepared/AssignPlaceholdersNoNorm-24     3.65MB ± 0%    3.65MB ± 0%     ~     (p=0.738 n=5+5)
Phases/many-columns-and-indexes-c/Prepared/AssignPlaceholdersNorm-24       3.65MB ± 0%    3.65MB ± 0%     ~     (p=0.841 n=5+5)
Phases/many-columns-and-indexes-c/Prepared/Explore-24                      25.2MB ± 0%    25.2MB ± 0%   -0.25%  (p=0.008 n=5+5)
Phases/many-columns-and-indexes-c/Prepared/ExecBuild-24                    25.4MB ± 0%    25.3MB ± 0%   -0.25%  (p=0.008 n=5+5)
Phases/batch-insert-one/Simple/Explore-24                                  63.4kB ± 0%    63.4kB ± 0%     ~     (p=0.190 n=5+5)
Phases/batch-insert-one/Simple/ExecBuild-24                                67.8kB ± 0%    67.8kB ± 0%     ~     (p=0.183 n=5+5)
Phases/batch-insert-one/Prepared/ExecBuild-24                              4.46kB ± 0%    4.46kB ± 0%     ~     (all equal)
Phases/batch-insert-many/Simple/Explore-24                                 3.69MB ± 0%    3.69MB ± 0%     ~     (p=0.548 n=5+5)
Phases/batch-insert-many/Simple/ExecBuild-24                               3.76MB ± 0%    3.76MB ± 0%     ~     (p=0.095 n=5+5)
Phases/batch-insert-many/Prepared/ExecBuild-24                             73.2kB ± 0%    73.2kB ± 0%     ~     (all equal)
Phases/ored-preds-100/Simple/Explore-24                                     979kB ± 0%     979kB ± 0%   -0.02%  (p=0.008 n=5+5)
Phases/ored-preds-100/Simple/ExecBuild-24                                  1.01MB ± 0%    1.01MB ± 0%   -0.02%  (p=0.008 n=5+5)
Phases/ored-preds-100/Prepared/ExecBuild-24                                29.6kB ± 0%    29.6kB ± 0%     ~     (all equal)
Phases/ored-preds-using-params-100/Simple/Explore-24                        962kB ± 0%     962kB ± 0%   -0.02%  (p=0.008 n=5+5)
Phases/ored-preds-using-params-100/Simple/ExecBuild-24                      991kB ± 0%     991kB ± 0%   -0.02%  (p=0.008 n=5+5)
Phases/ored-preds-using-params-100/Prepared/AssignPlaceholdersNoNorm-24     348kB ± 0%     348kB ± 0%     ~     (p=0.310 n=5+5)
Phases/ored-preds-using-params-100/Prepared/AssignPlaceholdersNorm-24       348kB ± 0%     348kB ± 0%     ~     (p=0.690 n=5+5)
Phases/ored-preds-using-params-100/Prepared/Explore-24                      742kB ± 0%     742kB ± 0%   -0.02%  (p=0.008 n=5+5)
Phases/ored-preds-using-params-100/Prepared/ExecBuild-24                    772kB ± 0%     772kB ± 0%   -0.02%  (p=0.008 n=5+5)

name                                                                     old allocs/op  new allocs/op  delta
Phases/kv-read/Simple/Explore-24                                             88.0 ± 0%      86.0 ± 0%   -2.27%  (p=0.008 n=5+5)
Phases/kv-read/Simple/ExecBuild-24                                           93.0 ± 0%      91.0 ± 0%   -2.15%  (p=0.008 n=5+5)
Phases/kv-read/Prepared/ExecBuild-24                                         9.00 ± 0%      9.00 ± 0%     ~     (all equal)
Phases/kv-read-const/Simple/Explore-24                                       88.0 ± 0%      86.0 ± 0%   -2.27%  (p=0.008 n=5+5)
Phases/kv-read-const/Simple/ExecBuild-24                                     93.0 ± 0%      91.0 ± 0%   -2.15%  (p=0.008 n=5+5)
Phases/kv-read-const/Prepared/ExecBuild-24                                   6.00 ± 0%      6.00 ± 0%     ~     (all equal)
Phases/tpcc-new-order/Simple/Explore-24                                       171 ± 0%       169 ± 0%   -1.17%  (p=0.008 n=5+5)
Phases/tpcc-new-order/Simple/ExecBuild-24                                     177 ± 0%       175 ± 0%   -1.13%  (p=0.008 n=5+5)
Phases/tpcc-new-order/Prepared/ExecBuild-24                                  9.00 ± 0%      9.00 ± 0%     ~     (all equal)
Phases/tpcc-delivery/Simple/Explore-24                                        168 ± 0%       166 ± 0%   -1.19%  (p=0.008 n=5+5)
Phases/tpcc-delivery/Simple/ExecBuild-24                                      174 ± 0%       172 ± 0%   -1.15%  (p=0.008 n=5+5)
Phases/tpcc-delivery/Prepared/AssignPlaceholdersNoNorm-24                    27.0 ± 0%      27.0 ± 0%     ~     (all equal)
Phases/tpcc-delivery/Prepared/AssignPlaceholdersNorm-24                      27.0 ± 0%      27.0 ± 0%     ~     (all equal)
Phases/tpcc-delivery/Prepared/Explore-24                                     77.0 ± 0%      75.0 ± 0%   -2.60%  (p=0.008 n=5+5)
Phases/tpcc-delivery/Prepared/ExecBuild-24                                   83.0 ± 0%      81.0 ± 0%   -2.41%  (p=0.008 n=5+5)
Phases/tpcc-stock-level/Simple/Explore-24                                     659 ± 0%       655 ± 0%   -0.61%  (p=0.008 n=5+5)
Phases/tpcc-stock-level/Simple/ExecBuild-24                                   680 ± 0%       676 ± 0%   -0.59%  (p=0.008 n=5+5)
Phases/tpcc-stock-level/Prepared/AssignPlaceholdersNoNorm-24                  106 ± 0%       106 ± 0%     ~     (all equal)
Phases/tpcc-stock-level/Prepared/AssignPlaceholdersNorm-24                    105 ± 0%       105 ± 0%     ~     (all equal)
Phases/tpcc-stock-level/Prepared/Explore-24                                   264 ± 0%       260 ± 0%   -1.52%  (p=0.008 n=5+5)
Phases/tpcc-stock-level/Prepared/ExecBuild-24                                 285 ± 0%       281 ± 0%   -1.40%  (p=0.008 n=5+5)
Phases/many-columns-and-indexes-a/Simple/Explore-24                           102 ± 0%        87 ± 0%  -14.71%  (p=0.008 n=5+5)
Phases/many-columns-and-indexes-a/Simple/ExecBuild-24                         111 ± 0%        96 ± 0%  -13.51%  (p=0.008 n=5+5)
Phases/many-columns-and-indexes-a/Prepared/AssignPlaceholdersNoNorm-24       21.0 ± 0%      21.0 ± 0%     ~     (all equal)
Phases/many-columns-and-indexes-a/Prepared/AssignPlaceholdersNorm-24         21.0 ± 0%      21.0 ± 0%     ~     (all equal)
Phases/many-columns-and-indexes-a/Prepared/Explore-24                        54.0 ± 0%      39.0 ± 0%  -27.78%  (p=0.008 n=5+5)
Phases/many-columns-and-indexes-a/Prepared/ExecBuild-24                      63.0 ± 0%      48.0 ± 0%  -23.81%  (p=0.008 n=5+5)
Phases/many-columns-and-indexes-b/Simple/Explore-24                           158 ± 0%       143 ± 0%   -9.24%  (p=0.008 n=5+5)
Phases/many-columns-and-indexes-b/Simple/ExecBuild-24                         171 ± 0%       157 ± 0%   -8.19%  (p=0.000 n=5+4)
Phases/many-columns-and-indexes-b/Prepared/AssignPlaceholdersNoNorm-24       36.0 ± 0%      36.0 ± 0%     ~     (all equal)
Phases/many-columns-and-indexes-b/Prepared/AssignPlaceholdersNorm-24         36.0 ± 0%      36.0 ± 0%     ~     (all equal)
Phases/many-columns-and-indexes-b/Prepared/Explore-24                        66.0 ± 0%      51.0 ± 0%  -22.73%  (p=0.008 n=5+5)
Phases/many-columns-and-indexes-b/Prepared/ExecBuild-24                      79.0 ± 0%      64.0 ± 0%  -18.99%  (p=0.008 n=5+5)
Phases/many-columns-and-indexes-c/Simple/Explore-24                          284k ± 0%      284k ± 0%   -0.02%  (p=0.008 n=5+5)
Phases/many-columns-and-indexes-c/Simple/ExecBuild-24                        284k ± 0%      284k ± 0%   -0.02%  (p=0.008 n=5+5)
Phases/many-columns-and-indexes-c/Prepared/AssignPlaceholdersNoNorm-24      56.4k ± 0%     56.4k ± 0%     ~     (all equal)
Phases/many-columns-and-indexes-c/Prepared/AssignPlaceholdersNorm-24        56.4k ± 0%     56.4k ± 0%     ~     (all equal)
Phases/many-columns-and-indexes-c/Prepared/Explore-24                        279k ± 0%      279k ± 0%   -0.02%  (p=0.008 n=5+5)
Phases/many-columns-and-indexes-c/Prepared/ExecBuild-24                      279k ± 0%      279k ± 0%   -0.02%  (p=0.008 n=5+5)
Phases/batch-insert-one/Simple/Explore-24                                     330 ± 0%       330 ± 0%     ~     (all equal)
Phases/batch-insert-one/Simple/ExecBuild-24                                   345 ± 0%       345 ± 0%     ~     (all equal)
Phases/batch-insert-one/Prepared/ExecBuild-24                                16.0 ± 0%      16.0 ± 0%     ~     (all equal)
Phases/batch-insert-many/Simple/Explore-24                                  29.4k ± 0%     29.4k ± 0%     ~     (p=0.492 n=5+5)
…